string strlen

+++ argc_argv/4-add.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 
 /**
  * main - multiply arguments
@@ -21,7 +22,7 @@ int main(int argc, char *argv[])
 	{
 		for (i = 0; i < argc; i++)
 		{
-			for (j = 0; j < strlen(argv[i]); j++)
+			for (j = 0; j < (int)strlen(argv[i]); j++)
 			{
 				if (!(argv[i][j] >= '0' && argv[i][j] <= '9'))
 				{
